+++ BE/src/main/java/com/team11/airbnb/web/controller/HomeController.java
@@ -1,6 +1,12 @@
 package com.team11.airbnb.web.controller;
 
-import com.team11.airbnb.openapi.DistrictSearchService;
+import java.util.List;
+
+import com.team11.airbnb.openapi.Position;
+import com.team11.airbnb.web.dto.AroundSpotDto;
+import com.team11.airbnb.web.dto.MainEventDto;
+import com.team11.airbnb.web.dto.ThemeSpotDto;
+import com.team11.airbnb.web.service.DistrictSearchService;
 import com.team11.airbnb.web.dto.HomeData;
 import com.team11.airbnb.web.service.HomeService;
 import lombok.RequiredArgsConstructor;
@@ -15,8 +21,11 @@ public class HomeController {
     private final DistrictSearchService districtSearchService;
 
     @GetMapping("/home")
-    public HomeData home() throws Exception {
-        return homeService.initApplication();
+    public HomeData home(Position position) throws Exception {
+        MainEventDto mainEventDto = homeService.getMainEventDto();
+        List<ThemeSpotDto> themeSpotDtoList = homeService.getThemeSpotDtoList();
+        List<AroundSpotDto> aroundSpotDtos = districtSearchService.searchDistrictInfo(position);
+        return new HomeData(mainEventDto, aroundSpotDtos, themeSpotDtoList);
     }
 
 }
